// frontend/src/shared/api/model/permissionTierEnum.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Student Evaluation System API
 * Outcome-based assessment system for academic programs
 * OpenAPI spec version: 1.0.0
 */

/**
 * * `view` - View Only
* `edit` - Edit
* `full` - Full Control
 */
export type PermissionTierEnum = typeof PermissionTierEnum[keyof typeof PermissionTierEnum];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const PermissionTierEnum = {
  view: 'view',
  edit: 'edit',
  full: 'full',
} as const;